Move Parameter definition to parameters.hpp
